change v2 into index and index into old index
