+++ src/pages/company/job-management/create-job.tsx
@@ -28,7 +28,7 @@ const CreateJob = () => {
                 Homepage
               </Typography>
             </Link>
-            <Link key='2' href='/company/job-management/v2' sx={{ textDecoration: 'none' }}>
+            <Link key='2' href='/company/job-management' sx={{ textDecoration: 'none' }}>
               <Typography
                 sx={{
                   color: '#32497A',
