+++ src/views/job-management/form/ProfessionalJob.tsx
@@ -277,7 +277,7 @@ const SeafarerJob = ({ job, type }: { job?: Job; type: 'create' | 'edit' }) => {
         .then(
           () => {
             toast.success(`${jobTitle} edited successfully!`)
-            router.push('/company/job-management/v2')
+            router.push('/company/job-management')
           },
           error => {
             toast.error('Failed to save job: ' + error.response.data.message)
@@ -290,7 +290,7 @@ const SeafarerJob = ({ job, type }: { job?: Job; type: 'create' | 'edit' }) => {
           () => {
             // clearDraft()
             toast.success(`${jobTitle} submited successfully!`)
-            router.push('/company/job-management/v2')
+            router.push('/company/job-management')
           },
           error => {
             toast.error('Failed to save job: ' + error.response.data.message)
